Added Adc16 Get Data Added @Andi comments

+++ Core/Inc/channels/adc24_channel.h
@@ -12,7 +12,7 @@ typedef struct
 } Adc24_Channel_t;
 
 Result_t Adc24_ProcessMessage(uint8_t ch_id, uint8_t cmd_id, uint8_t *data, uint32_t length);
-Result_t Adc24_GetData(uint8_t ch_id, int32_t *data, uint32_t *length);
+Result_t Adc24_GetData(uint8_t ch_id, int8_t *data, uint32_t *length);
 
 #endif
